# made FileManager store File and not open it each time